revise http stream storage
